IME: Added wxTextInputEvent class and wxEVT_TEXT_INPUT.
The definition of the class lives in a separate header file, because,
if I put it in <wx/event.h>, I would need to recompile irrelevant source files
whenever I change a single bit of wxTextInputEvent class.

The definition should be moved to <wx/event.h> once the API is stabilized.